launcher: one click component update (#35)
* change to vue3

without any errors

* new Update

it's a last dbug, with the better css

* Update OneClick.vue

page of component one click is created, with the style and template. scripts are not update

* new update

some new update with dbug

* new update for node page

node page added without content but in the route installed

* make ui

make a Trapezium on ui design and add this to the main js for use every wher

* update for the Node page

it's just simple page without any component. it has a small bug

* nodeBg added

the node page's back ground added to main.js. it has some css problems at he moment but it's ok. the routing at the moment is symple. without any conditions! 😏

* update oneclick with v-for

added v-for and data to the row but they are not inline block

* update

* update

* node header

a component added to header for the node page

* node header js

node header as a base component

* header added

the function is good but css has problem, positioning is not true