Add Log Navigation and Refactor Log Plugin Header Style (#1075)
* Refactor log header style

Simplified and improved the CSS styles for the log header section. Removed unnecessary styles and updated the layout for better consistency and readability. This refactor should make it easier to maintain and update the log header style in the future.

* Add(logs-plugin): Add navigation logs

Added navigation logs to the logs plugin to track user navigation events.

* Fix, log header style fixed with tiny css

* fix: scrollbar style fixed