ADD: Active Validator Check on Import (#1084)
* UPDATE: tutorial module words are dynamic

* Add Popup UI with Functions for Wire to Backend

This commit adds a new feature to our project, a popup user interface (UI) with functions for risk accept and cancel items. The popup UI is created to enhance the user experience by allowing users to perform actions quickly and efficiently without leaving the current page.

The new feature has been thoroughly tested and debugged to ensure its functionality. With this commit, the feature is ready to be wired to the backend.

* debug the staking css

* update the crowdin words

* update the functionallity the vertical bars and amsterdam comp. for sample

* tiny css update for the bars

* prettier format

* ADD: checking latest two Epoch

* update validator account

* update validator-check function

* update the checkvalidator in the frontend

* update function check validator

* fix the risk control modal

* FIX: removing tests

* fixes and smaller adjustments

* fix the cancel bug to show button again

* fix the error from dialog

* fix wrong args passed

---------

Co-authored-by: gbayasgalan <bayasgalan.gantumur@rocklogic.at>
Co-authored-by: NeoPlays <80448387+NeoPlays@users.noreply.github.com>